moved stopwatch to timings, sorted out the imports to be absolute relative.
