+++ tdash/__init__.py
@@ -1,5 +1,4 @@
-from functions import *
-from generators import *
-from timing import *
-from lru_cache import LRUCache
-from stopwatch import Stopwatch
+from .functions import *
+from .generators import *
+from .timing import *
+from .lru_cache import LRUCache
